chore: add and refactor endpoints

+++ src/deck/deck.service.ts
@@ -1,48 +1,52 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
+import { Prisma } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { handleError } from 'src/utils/handle-error.util';
 import { notFoundError } from 'src/utils/not-found.util';
 import { CreateDeckDto } from './dto/create-deck.dto';
-import { UpdateDeckDto } from './dto/update-deck.dto';
 
 @Injectable()
 export class DeckService {
   constructor(private readonly prisma: PrismaService) {}
-  async create(dto: CreateDeckDto) {
-    return await this.prisma.deck.create({ data: dto });
+  async create(dto: CreateDeckDto, loggedUserId) {
+    async function deckLimit(limitNumber) {
+      const limit = await this.prisma.deck.findMany({
+        where: { userId: loggedUserId },
+      });
+
+      if (limit.length >= limitNumber) {
+        throw new BadRequestException('Your deck has reached its limit.');
+      }
+    }
+
+    deckLimit(3);
+    const data: Prisma.DeckCreateInput = {
+      user: {
+        connect: {
+          id: loggedUserId,
+        },
+      },
+      userToCard: {
+        connect: {
+          id: dto.utcId,
+        },
+      },
+    };
+    return await this.prisma.deck.create({ data }).catch(handleError);
   }
 
-  async findOne(id: string) {
-    const data = await this.prisma.deck
-      .findUnique({ where: { id } })
-      .catch(handleError);
-
-    notFoundError(data, 'this deck');
-
-    return data;
-  }
-
-  async update(id: string, dto: UpdateDeckDto) {
-    const data = await this.prisma.deck
-      .update({ where: { id }, data: dto })
-      .catch(handleError);
-
+  async delete(id: string, userId: string) {
     notFoundError(
-      this.prisma.deck.findUnique({ where: { id } }),
+      this.prisma.deck.findUnique({ where: { id, userId } }),
       `this deck (${id})`,
     );
-
-    return data;
+    this.prisma.deck.delete({ where: { id, userId } }).catch(handleError);
+    return { message: 'Card sucessfuly removed.' };
   }
 
-  async delete(id: string) {
-    this.prisma.deck.delete({ where: { id } }).catch(handleError);
+  async reset(userId: string) {
+    this.prisma.deck.deleteMany({ where: { userId } });
 
-    notFoundError(
-      this.prisma.deck.findUnique({ where: { id } }),
-      `this deck (${id})`,
-    );
-
-    return { message: 'Sucessfully deleted' };
+    return { message: 'Your deck have been cleaned.' };
   }
 }
